Move install/upgrade tests to 'kudo' package
