+++ pkg/kudoctl/cmd/uninstall_test.go
@@ -6,10 +6,16 @@ import (
 	"github.com/kudobuilder/kudo/pkg/kudoctl/env"
 
 	"github.com/kudobuilder/kudo/pkg/apis/kudo/v1alpha1"
+	"github.com/kudobuilder/kudo/pkg/client/clientset/versioned/fake"
+	"github.com/kudobuilder/kudo/pkg/kudoctl/util/kudo"
 	util "github.com/kudobuilder/kudo/pkg/util/kudo"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
+
+func newTestClient() *kudo.Client {
+	return kudo.NewClientFromK8s(fake.NewSimpleClientset())
+}
 
 func TestUninstall(t *testing.T) {
 	testInstance := v1alpha1.Instance{
